requirements.txt
gpiozero
paho-mqtt
psutil
RPi.GPIO
